requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
openai>=1.51.0
Pillow==10.1.0
PyQt5==5.15.10
pyttsx3==2.90
pyobjc-framework-Quartz==10.1
pyobjc-framework-ApplicationServices==10.1
python-dotenv==1.0.0